# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SecureProfit Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/tasks:
    get:
      operationId: listProjectTasks
      tags: [tasks]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Task" }
    post:
      operationId: createProjectTask
      tags: [tasks]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTaskBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
...
  /tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [tasks]
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateTaskBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
    delete:
      operationId: deleteTask
      tags: [tasks]
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessMessage" }
  /tasks/{taskId}/time-logs:
    get:
      operationId: listTaskTimeLogs
      tags: [tasks]
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TaskTimeLog" }
    post:
      operationId: logTaskTime
      tags: [tasks]
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LogTaskTimeBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskTimeLog" }
...
components:
  schemas:
...
    SuccessMessage:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        message: { type: string }
...
    TaskStatus:
      type: string
      enum: [TODO, IN_PROGRESS, BLOCKED, DONE]
...
    TaskAssignee:
      type: object
      required: [userId, name]
      properties:
        userId: { type: string }
        name: { type: string }
...
    TaskDependencyRef:
      type: object
      required: [id, dependsOnTaskId]
      properties:
        id: { type: string }
        dependsOnTaskId: { type: string }
        dependsOnTitle: { type: string, nullable: true }
    Task:
      type: object
      required: [id, projectId, title, status, progressPercent, billable, loggedHours, assignees, createdAt, updatedAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        workstreamId: { type: string, nullable: true }
        projectCode: { type: string, nullable: true }
        projectName: { type: string, nullable: true }
        title: { type: string }
        description: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/TaskStatus" }
        progressPercent: { type: integer, minimum: 0, maximum: 100 }
        billable: { type: boolean }
        plannedHours: { type: number, nullable: true, description: "Hour cap for this task (all users, non-rejected timesheets); null = uncapped" }
        timesheetHours: { type: number, description: "Total non-rejected timesheet hours clocked against this task" }
        startDate: { type: string, nullable: true }
        endDate: { type: string, nullable: true }
        assigneeId: { type: string, nullable: true }
        assigneeName: { type: string, nullable: true }
        assignees:
          type: array
          items: { $ref: "#/components/schemas/TaskAssignee" }
        parentTaskId: { type: string, nullable: true }
        subtaskCount: { type: integer }
        dependencies:
          type: array
          items: { $ref: "#/components/schemas/TaskDependencyRef" }
        createdById: { type: string, nullable: true }
        createdByName: { type: string, nullable: true }
        loggedHours: { type: number }
        createdAt: { type: string }
        updatedAt: { type: string }
    CreateTaskBody:
      type: object
      required: [title]
      properties:
        title: { type: string }
        description: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/TaskStatus" }
        progressPercent: { type: integer, minimum: 0, maximum: 100 }
        billable: { type: boolean }
        plannedHours: { type: number, nullable: true }
        startDate: { type: string, nullable: true }
        endDate: { type: string, nullable: true }
        assigneeId: { type: string, nullable: true }
        assigneeIds:
          type: array
          items: { type: string }
        parentTaskId: { type: string, nullable: true }
        dependencyTaskIds:
          type: array
          items: { type: string }
        workstreamId: { type: string, nullable: true }
    UpdateTaskBody:
      type: object
      properties:
        title: { type: string }
        description: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/TaskStatus" }
        progressPercent: { type: integer, minimum: 0, maximum: 100 }
        billable: { type: boolean }
        plannedHours: { type: number, nullable: true }
        startDate: { type: string, nullable: true }
        endDate: { type: string, nullable: true }
        assigneeId: { type: string, nullable: true }
        assigneeIds:
          type: array
          items: { type: string }
        parentTaskId: { type: string, nullable: true }
        dependencyTaskIds:
          type: array
          items: { type: string }
        workstreamId: { type: string, nullable: true }
    TaskTimeLog:
      type: object
      required: [id, taskId, userId, hours, loggedAt, createdAt]
      properties:
        id: { type: string }
        taskId: { type: string }
        userId: { type: string }
        userName: { type: string, nullable: true }
        hours: { type: number }
        note: { type: string, nullable: true }
        loggedAt: { type: string }
        createdAt: { type: string }
    LogTaskTimeBody:
      type: object
      required: [hours]
      properties:
        hours: { type: number }
        note: { type: string }
        loggedAt: { type: string }
...