# lib/api-zod/src/generated/types/updateTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SecureProfit Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskStatus } from "./taskStatus";

export interface UpdateTaskBody {
  title?: string;
  description?: string | null;
  status?: TaskStatus;
  /**
   * @minimum 0
   * @maximum 100
   */
  progressPercent?: number;
  billable?: boolean;
  plannedHours?: number | null;
  startDate?: string | null;
  endDate?: string | null;
  assigneeId?: string | null;
  assigneeIds?: string[];
  parentTaskId?: string | null;
  dependencyTaskIds?: string[];
  workstreamId?: string | null;
}